start refactor to plugins
