--- ne_engine/ne_graphics/public/src/Window.cpp
@@ -1,4 +1,0 @@
-#include "Window.hpp"
-
-
-
